Refactor Hexdump::Client to allow simpler extension
I'm working on a sequel-hexspace adapter, which requires:

* Column names must be symbols

* timestamp types need to be converted using a Sequel method that
  handles configured timezones

* binary types need to be converted to a Sequel-specific type

* Columns in a result set should be accessible even if the result
  set has no rows.

This is currently only possible by subclassing Hexspace::Client
and overriding process_results, which is quite a large method.
This extracts the following methods from process_results:

* column_names(metadata)

* type_converter(type)

It adds one keyword argument to the initializer for whether
process_results should include columns as a second return
value.

This makes the overriding sequel-hexspace needs to do much simpler:
https://github.com/jeremyevans/sequel-hexspace/commit/e70efc8b75e7320598a2ac0e7b6d3b05edb9327e